refactor(models) split off `tasks`
